+++ Citadel/src/utopia/citadel/database/access/single/language/DbLanguage.scala
@@ -1,7 +1,9 @@
 package utopia.citadel.database.access.single.language
 
 import utopia.citadel.database.access.id.single.DbLanguageId
+import utopia.citadel.database.access.many.description.DbDescriptions
 import utopia.citadel.database.factory.language.LanguageFactory
+import utopia.citadel.database.model.language.LanguageModel
 import utopia.flow.datastructure.immutable.Pair
 import utopia.flow.generic.ValueConversions._
 import utopia.flow.util.CollectionExtensions._
@@ -9,7 +11,10 @@ import utopia.metropolis.model.error.NoDataFoundException
 import utopia.metropolis.model.post.NewLanguageProficiency
 import utopia.metropolis.model.stored.language.Language
 import utopia.vault.database.Connection
-import utopia.vault.nosql.access.single.model.SingleModelAccessById
+import utopia.vault.nosql.access.single.model.distinct.{SingleIdModelAccess, UniqueModelAccess}
+import utopia.vault.nosql.access.single.model.SingleRowModelAccess
+import utopia.vault.nosql.template.Indexed
+import utopia.vault.nosql.view.{SubView, UnconditionalView}
 
 import scala.util.{Failure, Success, Try}
 
@@ -18,16 +23,32 @@ import scala.util.{Failure, Success, Try}
   * @author Mikko Hilpinen
   * @since 2.5.2020, v1.0
   */
-object DbLanguage extends SingleModelAccessById[Language, Int]
+object DbLanguage extends SingleRowModelAccess[Language] with UnconditionalView with Indexed
 {
-	// IMPLEMENTED	--------------------------------
+	// COMPUTED ------------------------------------
 	
-	override def idToValue(id: Int) = id
+	private def model = LanguageModel
+	
+	
+	// IMPLEMENTED	--------------------------------
 	
 	override def factory = LanguageFactory
 	
 	
 	// OTHER	------------------------------------
+	
+	/**
+	 * @param id A language id
+	 * @return An access point to that language
+	 */
+	def apply(id: Int) = new DbSingleLanguage(id)
+	
+	/**
+	 * @param languageCode A language code
+	 * @param connection Implicit DB Connection
+	 * @return An access point to a language with that ISO-code
+	 */
+	def forIsoCode(languageCode: String)(implicit connection: Connection) = new DbLanguageForIsoCode(languageCode)
 	
 	/**
 	  * Validates the proposed language proficiencies, making sure all language ids and codes are valid
@@ -59,5 +80,52 @@ object DbLanguage extends SingleModelAccessById[Language, Int]
 					Failure(new NoDataFoundException(s"${ proficiency.familiarityId } is not a valid language familiarity id"))
 			}
 		}
+	}
+	
+	
+	// NESTED   -----------------------------------
+	
+	class DbSingleLanguage(id: Int) extends SingleIdModelAccess[Language](id, factory)
+	{
+		/**
+		 * @param connection Implicit DB Connection
+		 * @return The ISO-code associated with this language
+		 */
+		def isoCode(implicit connection: Connection) = pullAttribute(model.isoCodeAttName)
+		
+		/**
+		 * @return An access point to this language's descriptions
+		 */
+		def descriptions = DbDescriptions.ofLanguageWithId(id)
+	}
+	
+	class DbLanguageForIsoCode(val code: String) extends UniqueModelAccess[Language] with SubView
+	{
+		// COMPUTED -------------------------------
+		
+		/**
+		 * @param connection Implicit DB Connection
+		 * @return Id of this language, if found
+		 */
+		def id(implicit connection: Connection) = pullColumn(index).int
+		
+		
+		// IMPLEMENTED  ---------------------------
+		
+		override protected def parent = DbLanguage
+		
+		override def filterCondition = model.withIsoCode(code).toCondition
+		
+		override def factory = parent.factory
+		
+		
+		// OTHER    --------------------------------
+		
+		/**
+		 * Reads this language, inserting one if it doesn't exist already
+		 * @param connection Implicit DB Connection
+		 * @return Read or inserted language
+		 */
+		def getOrInsert()(implicit connection: Connection) = pull.getOrElse { model.insert(code) }
 	}
 }